+++ compiler/src/sct_checker_forward.ml
@@ -751,7 +751,9 @@ let expr_equal a b =
     let open Glob_options in
     match !target_arch with
     | X86_64 -> X86_decl.x86_fcp
-    | ARM_M4 -> Arm_decl.arm_fcp in
+    | ARM_M4 -> Arm_decl.arm_fcp 
+    | RISCV  -> Riscv_decl.riscv_fcp
+  in 
   let normalize e =
     e |> Conv.cexpr_of_expr |> Constant_prop.(const_prop_e fcp None empty_cpm) in
   Expr.eq_expr (normalize a) (normalize b)
@@ -1311,12 +1313,10 @@ let init_constraint fenv f =
       | None ->
         begin match x.v_kind with
         | Const -> Env.dpublic env
-        | Stack Direct when is_local -> Direct (Env.fresh env, Env.secret env)
-        | Stack Direct -> Direct (Env.fresh2 env)
-        | Stack (Pointer _) when is_local -> Indirect((Env.fresh env, Env.secret env), Env.fresh2 env)
-        | Stack (Pointer _) -> Indirect(Env.fresh2 env, Env.fresh2 env)
-        | Reg (_, Direct) -> Direct (Env.fresh2 env)
-        | Reg (_, Pointer _) -> Indirect(Env.fresh2 env, Env.fresh2 env)
+        | (Stack Direct | Reg (_, Direct)) when is_local -> Direct (Env.secret2 env)
+        | (Stack Direct | Reg (_, Direct)) -> Direct (Env.fresh2 env)
+        | (Stack (Pointer _) | Reg (_, Pointer _)) when is_local -> Indirect(Env.secret2 env, Env.fresh2 env)
+        | (Stack (Pointer _) | Reg (_, Pointer _)) -> Indirect(Env.fresh2 env, Env.fresh2 env)
         | Inline -> Env.dpublic env
         | Global -> Env.dpublic env (* unsure *)
         end