+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 
 ## New features
 
+- Support RISC-V 32IM as target architecture. This comes with two new passes
+  dedicated for RISC-V. The “lower complex addressing mode” pass introduces
+  two auxiliary instructions for complex memory accesses not compilable as is
+  on RISC-V. The “load constants in conditions” pass introduces auxiliary
+  registers for non-zero constants appearing in conditions.
+  ([PR #781](https://github.com/jasmin-lang/jasmin/pull/781);
+  fixes [#503](https://github.com/jasmin-lang/jasmin/issues/503),
+  [#698](https://github.com/jasmin-lang/jasmin/issues/698)).
+
 - ARM now emits two instructions instead of `ADR` to load the low and high
   parts of global addresses
   ([PR#921](https://github.com/jasmin-lang/jasmin/pull/921)).
@@ -17,8 +26,6 @@
   ([PR #969](https://github.com/jasmin-lang/jasmin/pull/969);
   fixes [#902](https://github.com/jasmin-lang/jasmin/issues/902)).
 
-
-
 ## Bug fixes
 
 - Fix EasyCrypt semantics of shift operators
@@ -26,6 +33,11 @@
 
 - Fix EC extraction in case on nested loops
   ([PR #971](https://github.com/jasmin-lang/jasmin/pull/971)).
+
+- The speculative constant-time checker rejects more programs, considering that
+  different source-level local variables may be merged during the compilation
+  process
+  ([PR#989](https://github.com/jasmin-lang/jasmin/pull/989)).
 
 ## Other changes